migrate bytecode
