+++ bsn/rollup-finality-provider/config/config.go
@@ -10,8 +10,8 @@ import (
 
 type OPStackL2Config struct {
 	OPStackL2RPCAddress      string `long:"opstackl2-rpc-address" description:"the rpc address of the op-stack-l2 node to connect to"`
-	OPFinalityGadgetAddress  string `long:"op-finality-gadget" description:"the contract address of the op-finality-gadget"`
 	BabylonFinalityGadgetRpc string `long:"babylon-finality-gadget-rpc" description:"the rpc address of babylon op finality gadget"` //nolint:stylecheck,revive
+	FinalityContractAddress  string `long:"finality-contract-address" description:"the contract address of the rollup finality contract"`
 
 	// Below configurations are needed for the Babylon client
 	Common *fpcfg.Config
@@ -21,7 +21,7 @@ func (cfg *OPStackL2Config) Validate() error {
 	if cfg.OPStackL2RPCAddress == "" {
 		return fmt.Errorf("opstackl2-rpc-address is required")
 	}
-	_, _, err := bech32.Decode(cfg.OPFinalityGadgetAddress, len(cfg.OPFinalityGadgetAddress))
+	_, _, err := bech32.Decode(cfg.FinalityContractAddress, len(cfg.FinalityContractAddress))
 	if err != nil {
 		return fmt.Errorf("op-finality-gadget: invalid bech32 address: %w", err)
 	}
